yay completed

+++ src/app/api/menu-items/route.js
@@ -1,25 +1,24 @@
-import { isAdmin } from "@/app/api/auth/[...nextauth]/route";
-import { MenuItem } from "@/models/MenuItem";
+import {isAdmin} from "@/app/api/auth/[...nextauth]/route";
+import {MenuItem} from "@/models/MenuItem";
 import mongoose from "mongoose";
 
 export async function POST(req) {
-    mongoose.connect(process.env.MONGO_URL);
-    const data = await req.json();
+  mongoose.connect(process.env.MONGO_URL);
+  const data = await req.json();
+  if (await isAdmin()) {
     const menuItemDoc = await MenuItem.create(data);
-    //   if (await isAdmin()) {
-    //     return Response.json(menuItemDoc);
-    //   } else {
-    //     return Response.json({});
-    //   }
-    return Response.json(menuItemDoc)
+    return Response.json(menuItemDoc);
+  } else {
+    return Response.json({});
+  }
 }
 
 export async function PUT(req) {
   mongoose.connect(process.env.MONGO_URL);
-  // if (await isAdmin()) {
+  if (await isAdmin()) {
     const {_id, ...data} = await req.json();
     await MenuItem.findByIdAndUpdate(_id, data);
-  // }
+  }
   return Response.json(true);
 }
 
@@ -34,8 +33,8 @@ export async function DELETE(req) {
   mongoose.connect(process.env.MONGO_URL);
   const url = new URL(req.url);
   const _id = url.searchParams.get('_id');
-  // if (await isAdmin()) {
+  if (await isAdmin()) {
     await MenuItem.deleteOne({_id});
-  // }
+  }
   return Response.json(true);
 }
